requirements.txt
Flask==2.2.2
langchain==0.0.163
groq==0.1.3
reportlab==3.6.1
PyMuPDF==1.21.1
json5==0.9.6
tenacity==8.0.1
python-dotenv==1.0.0
PyPDF2==2.11.0
transformers==4.28.1
speechrecognition==3.8.1
numpy==1.23.5
Werkzeug==2.2.2
opencv-python==4.6.0.66
pytesseract==0.3.9
Pillow==9.3.0
fpdf==1.7.2
ultralytics==8.0.24
